test(webview): fold MainViewActions suite into the launch-target suite
Both files tested mainViewActions/mainViewState through byte-identical
hoisted hostBridge mocks and duplicate reset scaffolding. The direct-
actions describe moves verbatim (all 15 cases) into the merged file,
renamed MainViewActions.vitest.ts to path-mirror the module it owns.
